+++ src/test/java/bowling/domain/game/FrameNumberTest.java
@@ -1,4 +1,4 @@
-package bowling.game.domain;
+package bowling.domain.game;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
